Move tests specific functionalities from libvmi to the tests
