+++ tests/libvmi/BUILD.bazel
@@ -16,14 +16,10 @@ go_library(
         "//pkg/controller:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
-        "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//tests/containerdisk:go_default_library",
         "//tests/framework/checks:go_default_library",
-        "//tests/framework/matcher:go_default_library",
         "//tests/libnet:go_default_library",
-        "//tests/libstorage:go_default_library",
         "//tests/testsuite:go_default_library",
-        "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
